// apps/web/src/api/generated/organizer-tournaments/organizerTournamentsControllerListOrganizerTournamentsSortBy.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type OrganizerTournamentsControllerListOrganizerTournamentsSortBy = typeof OrganizerTournamentsControllerListOrganizerTournamentsSortBy[keyof typeof OrganizerTournamentsControllerListOrganizerTournamentsSortBy];


export const OrganizerTournamentsControllerListOrganizerTournamentsSortBy = {
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
  startsAt: 'startsAt',
  registrationClosesAt: 'registrationClosesAt',
  name: 'name',
} as const;