// apps/web/src/api/generated/organizer-tournaments/organizerTournamentsControllerListOrganizerTournamentsSortDirection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type OrganizerTournamentsControllerListOrganizerTournamentsSortDirection = typeof OrganizerTournamentsControllerListOrganizerTournamentsSortDirection[keyof typeof OrganizerTournamentsControllerListOrganizerTournamentsSortDirection];


export const OrganizerTournamentsControllerListOrganizerTournamentsSortDirection = {
  asc: 'asc',
  desc: 'desc',
} as const;